객체 생성, 및 회전 적용.

+++ WinAPI/CImageUI.cpp
@@ -62,7 +62,7 @@ void CImageUI::Render(HDC _hDC)
     }
     else { // Texture가 있으면
         CTexture* pTexture = GetTexture();
-        BitBlt(_hDC, (int)vFinalPos.x, (int)vFinalPos.y, GetScale().x, GetScale().y, pTexture->GetDC(), 1, 1, SRCCOPY);
+        BitBlt(_hDC, (int)vFinalPos.x, (int)vFinalPos.y, (int)GetScale().x, (int)GetScale().y, pTexture->GetDC(), 1, 1, SRCCOPY);
         //BitBlt(_hDC, (int)vFinalPos.x, (int)vFinalPos.y, g_iTileSize, g_iTileSize, pTexture->GetDC(), 1,1, SRCCOPY);
     }
